+++ src/main/java/com/middle/advanced/service/UserService.java
@@ -34,7 +34,7 @@ public class UserService {
     }
 
     public TUser selectUserById(Long id) {
-        return userMapper.getOne(id.intValue());
+        return userMapper.getOne(id);
     }
 
     public String login(LoginDto login) {
